Create and register root blueprint in controllers with one path
Add a single controller that returns a json message when a GET request 
is sent to '/'. This blueprint is registered using a pattern I haven't 
seen before that I got from Dylan Anthony's tutorial; This pattern 
registers all blueprints in a single function in 
controllers/__init__.py, and this function is in turn called from the 
main application factory. One thing I like about this pattern is that it 
matches the way that other Flask extensions are registered, though it 
may be good to have registration for extensions look syntactically 
different from registering things within the app.

Pytest's single test now passes.