Install black and safety with pipenv
Black is an up-and-coming style enforcer that was recomended by the
tutorial I am folllowing. I do not use a style checker yet, but want to
start using one. This one looks nice, and it describes itself as a
'strict subset of PEP-8', which I like, coming from JS where a lot of
linters were very opinionated and conflicted with eachother.

Safety is a command line tool that checks for security vulnerabilities
in your dependencies; this sounds like a very good thing to have as a CI
tool.