Add mock json data to get_all_articles response
Added a hard-coded dictionary of three mock articles to the articles
controller, which is returned as JSON whenever a GET request is made to
the root of articles. Also added this same hard-coded dictionary to the
unit test for this endpoint. I plan to first create all my endpoints
using in-memory data, then refactor it to use a database. I hope that by
doing it this way I will be forced to decouple the database interactions
from the API logic.