Add configuration
Added configuration from classes to application factory. Config classes
are in a file named config.py, which is in the project root directory.
There are three configurations - development, testing, and production -
which inherit from a base configuration class. Right now the
configuration classes themselves are rather sparse.

Configuration can be specified explicitly as an argument to create_app,
or can be set using the FLASK_ENV environment variable, which is the way
that configuration is set in the Flask Documentation in many places.
Flask doesn't have a lot of set-in-stone best practices due to its
nature as a microframework, but I assume that anything shown in the
documentation is a generally correct way to do something.