Add flask-sqlalchemy and config for sqlite database
Add flask-sqlalchemy to application factory in __init__.py, which is initialized within a helper function. This pattern is used so that everything stays organized when we start adding more packages.

Added database path to configuration files. Databases are sqlite databases stored in the project root, but not in version control. For projects at scale I would use a psql database, but for this project sqlite is the better choice due to its simplicity.