Fix database paths, add database creation to app factory
I learned that database paths in the config file are relative to the api/ directory, and not the config file itself. I have updated the paths accordingly to place the database files in the project root directory.

I added a line within the application factory to create all database tables