Install slugify, refactor article model, add endpoint for single article
Install slugify to format slugs for URLs
Refactor articles model to have slug column, and to initialize the slug column as a slugified version of the title column. At least for now it makes sense to me to put the slugify function within the article model.
I am pondering how much logic should be within the model vs. the controller. Should slugification be done in the model, or should all logic be in the controller? I'm leaning towards the latter and will likely refactor the articles model and controller to reflect that soon.