Add error handling for articles POST: missing or incomplete request body
Add error handling for POST to /articles/ that will return an error message if the request body is missing or incomplete. There is more error handling to do - such as if an article is created with a title that will create a slug that is already in use - but I have to think about how to solve that one before I write a test for it.

I plan to refactor my request validation into some function or method once all error handling is implemented.