Add failing test: create article w/ same slug/title as existing article
Add a test to check whether a POSTed article has a title that will,when slugified, match an existing article's slug.

The easiest way to pass this test would be to import slugify into the articles controller, create a slug from the request title, and then query the database by slug to see what comes back. This solution is brittle and doesn't scale:
* Am I going to query the database to validate every attribute of my model? What about for models with many attributes? That adds up.
* I don't like that I am now generating slugs in two different places, and I don't like that I would be essentially mocking my model so as to validate it. This is brittle because if the model ever changes the mock must also be updated. This is also conceptually wrong because I would be validating the mock, rather than validating the model itself, which is what I really want to do.

The best solution I can conceive of is to have the validation within the model - as a method or methods: Maybe a try/except block within the constructor? Maybe a validator method that is called separately? What would be ideal would be to have validation done during initialization of the model: If the data being passed into the model is invalid, it spits out an error instead of creating an instance of the model.

Some quick researching reveals that adding exception handling to a class constructor is not a trivial or obvious matter, so I will need to do some more research into how to handle this.