Add type annotations to test_models.py
Added type annotations to test_models.py. Even though I don't use mypy
on tests, I wanted to annotate these tests in particular for
documentation purposes. It was a major bugger to get serialization and
deserialization working due to TypeErrors and not knowing what type my
inputs and outputs needed to be, and what types to comparte them
against.