Add validation for existing article titles
Add validation method to article schema to check that slug will be unique. The way the validation method is implemented is the only way I found to do it while still initializing the slug in the constructor, rather than in a marshmallow pre_load method.
---
Very good article on validating in flask-marshmallow:
https://www.cameronmacleod.com/blog/better-validation-flask-marshmallow