Refactor slug validator to pass test
Slug validator should now work correctly for creating new articles as well as updating existing articles:

If creating a new article, the slug should always be unique. New articles are identified by lacking an id, since the id is generated in the database.
If updating an existing article(has id), then slug only needs to be checked for uniqueness if it has been changed. If an articles is being updated and the slug hasn't changed, then the article found in the query is this very article.

I will be writing some tests to verify that the logic here works correctly.