Add simple list of available endpoints to root.
Add a simple list of available endpoints to the root path("/"). Now the root path of the application does not return a 404 error.

In a full-blown restful api, perhaps this would be complete documentation, including accepted methods, request parameters, response data, etc.
But I am questioning the scope of this project, and I don't think total documentation is within the scope.