Update readme
Remove development notes - you can view them in the commit history.

Added a bit of an introduction about the project - reduced the scope of
the project from "build the best api in the best way with all the
things", which was absurdly over-ambitious. I learned a lot from this
project, but I also learned how much I don't know.